# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Clarity B2B API",
    "version": "1.0.0",
    "description": "API documentation for the Clarity Supply Chain Finance backend. This covers the Invoice Pipeline and Supplier Health Analytics modules."
  },
  "servers": [
    {
      "url": "http://localhost:5001",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/api/disputes": {
      "get": {
        "summary": "Get all disputes",
        "description": "Get a list of all filed disputes.",
        "responses": {
          "200": {
            "description": "A JSON array of disputes"
          }
        }
      },
      "post": {
        "summary": "File a new dispute",
        "description": "File a dispute against an invoice, freezing it.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "invoice_id": {
                    "type": "string"
                  },
                  "filed_by": {
                    "type": "string"
                  },
                  "reason": {
                    "type": "string"
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Dispute filed successfully"
          }
        }
      }
    },
    "/api/disputes/{id}": {
      "get": {
        "summary": "Get dispute details",
        "description": "Get the details of a specific dispute.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dispute details"
          },
          "404": {
            "description": "Dispute not found"
          }
        }
      }
    },
    "/api/disputes/{id}/evidence": {
      "post": {
        "summary": "Add evidence to a dispute",
        "description": "Upload a document link to support an open dispute.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "uploaded_by": {
                    "type": "string"
                  },
                  "file_url": {
                    "type": "string"
                  },
                  "note": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Evidence added successfully"
          },
          "404": {
            "description": "Dispute not found"
          }
        }
      }
    },
    "/api/disputes/{id}/resolve": {
      "patch": {
        "summary": "Resolve a dispute",
        "description": "Release the invoice back to the marketplace or void it entirely.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "decision": {
                    "type": "string",
                    "enum": ["released", "voided"]
                  },
                  "actor": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Dispute resolved successfully"
          },
          "404": {
            "description": "Dispute not found"
          }
        }
      }
    },
...
  }
}